Move OpenUvConfigEntry to coordinator module
