+++ homeassistant/components/openuv/coordinator.py
@@ -20,11 +20,13 @@ from .const import LOGGER
 
 DEFAULT_DEBOUNCER_COOLDOWN_SECONDS = 15 * 60
 
+type OpenUvConfigEntry = ConfigEntry[dict[str, OpenUvCoordinator]]
+
 
 class OpenUvCoordinator(DataUpdateCoordinator[dict[str, Any]]):
     """Define an OpenUV data coordinator."""
 
-    config_entry: ConfigEntry
+    config_entry: OpenUvConfigEntry
     update_method: Callable[[], Awaitable[dict[str, Any]]]
 
     def __init__(
